移除 str(x) 宏定义, 气数、攻击数、防御数全部使用整型, 添加 gen_mapx 以更快生成整型, 添加 formatx, formatxstr 以更快输出整型, Step 2 补充招式计数, Step 5 修正了技能防御添加的逻辑, Step 7, Step 9 添加防御值为-1(无限)的特判, Step 11 添加, Step 12 补充镐类等值爆

+++ wooden_status.cpp
@@ -21,7 +21,7 @@ game_status::game_status(int _player_num) {
 }
 
 game_status::game_status(int& _player_num, const std::vector<int>* _players,
-                         const std::map<int, float>& _qi,
+                         const std::map<int, int>& _qi,
                          const std::map<int, bool>& _tag_died,
                          const std::map<int, std::map<int, int> >& _skl_count) {
     player_num = _player_num;
